# speedray-demo-resource/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "1.0.0",
    "title" : "Liferay DXP JAX RS Services using Xtivia Speedray Publisher"
  },
  "host" : "localhost:8080",
  "basePath" : "/o/rest/samples",
  "schemes" : [ "http" ],
  "paths" : {
    "/people" : {
      "get" : {
        "tags" : [ "/samples" ],
        "summary" : "Returns the entire collection of in-memory Person objects",
        "description" : "",
        "operationId" : "getAllPeople",
        "produces" : [ "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Person"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "/samples" ],
        "summary" : "Creates a new Person object based on suppplied JSON and returns the created object including ID",
        "description" : "",
        "operationId" : "addPerson",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Person"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Person"
            }
          }
        }
      }
    },
    "/people/{id}" : {
      "get" : {
        "tags" : [ "/samples" ],
        "summary" : "Returns a single Person object based on suppplied ID",
        "description" : "",
        "operationId" : "getPerson",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Person"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "/samples" ],
        "summary" : "Modifies an existing Person object based on suppplied JSON and uses the ID parameter to identify the target person",
        "description" : "",
        "operationId" : "updatePerson",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Person"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Person"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "/samples" ],
        "summary" : "Removes a single Person object based on suppplied ID",
        "description" : "",
        "operationId" : "deletePerson",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Person"
            }
          }
        }
      }
    },
    "/users/current" : {
      "get" : {
        "tags" : [ "/samples" ],
        "summary" : "Returns a simplified JSON representation of Liferay User object for logged in user (or default user if not logged in)",
        "description" : "",
        "operationId" : "getCurrentUser",
        "produces" : [ "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/LiferayUser"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "LiferayUser" : {
      "type" : "object",
      "properties" : {
        "emailAddress" : {
          "type" : "string"
        },
        "screenName" : {
          "type" : "string"
        },
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "active" : {
          "type" : "boolean",
          "default" : false
        },
        "defaultUser" : {
          "type" : "boolean",
          "default" : false
        }
      }
    },
    "Person" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "lastName" : {
          "type" : "string"
        },
        "firstName" : {
          "type" : "string"
        },
        "location" : {
          "type" : "string"
        }
      }
    }
  }
}